Fix lazy and refactor
